+++ R/sim_trait.R
@@ -31,6 +31,10 @@
 #' Note that this threshold is applied to the sample allele frequencies and not their true parametric values (\code{p_anc}), even if these are available.
 #' @param loci_on_cols If \code{TRUE}, \eqn{X} has loci on columns and individuals on rows; if false (the default), loci are on rows and individuals on columns.
 #' If \eqn{X} is a BEDMatrix object, loci are taken to be on the columns (regardless of the value of \code{loci_on_cols}).
+#' @param mem_factor BEDMatrix-specific, sets proportion of available memory to use loading genotypes.
+#' Ignored if `mem_lim` is not `NA`.
+#' @param mem_lim BEDMatrix-specific, sets total memory to use loading genotypes, in GB.
+#' If `NA` (default), a proportion `mem_factor` of the available memory will be used.
 #'
 #' @return A list containing the simulated \code{trait} (length \eqn{n}), the vector of causal locus indexes \code{causal_indexes} (length \eqn{m_causal}), and the locus effect size vector \code{causal_coeffs} (length \eqn{m_causal}) at the causal loci.
 #' However, if \code{herit = 0} then \code{causal_indexes} and \code{causal_coeffs} will have zero length regardless of \code{m_causal}.
@@ -56,7 +60,7 @@
 #' obj$causal_coeffs
 #' 
 #' @export
-sim_trait <- function(X, m_causal, herit, p_anc, kinship, mu = 0, sigma_sq = 1, maf_cut = 0.05, loci_on_cols = FALSE) {
+sim_trait <- function(X, m_causal, herit, p_anc, kinship, mu = 0, sigma_sq = 1, maf_cut = 0.05, loci_on_cols = FALSE, mem_factor = 0.7, mem_lim = NA) {
     # check for missing parameters
     if (missing(X))
         stop('genotype matrix `X` is required!')
@@ -103,7 +107,7 @@ sim_trait <- function(X, m_causal, herit, p_anc, kinship, mu = 0, sigma_sq = 1, 
         ###################
         
         # compute marginal allele frequencies
-        p_anc_hat <- allele_freqs(X, loci_on_cols = loci_on_cols)
+        p_anc_hat <- allele_freqs(X, loci_on_cols = loci_on_cols, mem_factor = mem_factor, mem_lim = mem_lim)
         
         # select random SNPs! this performs the magic...
         # also runs additional checks